% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box_plot.R
\name{Boxplot_trans}
\alias{Boxplot_trans}
\title{Internal Use Function}
\usage{
Boxplot_trans(methallchr)
}
\arguments{
\item{methallchr}{refers to the filtered subset of inputmethfile.}
}
\value{
Outputs dataframe.
}
\description{
This function transforms filtered subset of inputmethfile to a dataframe for boxplot.
}
\examples{
Boxplot_trans(methallchr)

}
