% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pie_plot.R
\name{Chromosome_pieplot}
\alias{Chromosome_pieplot}
\title{Pie plot based on different chromosomes.}
\usage{
Chromosome_pieplot(genefeatureall_cpgfeature_significantcpgisland,
  genefeatureall_cpgfeature_significantshore = NULL,
  methdirection = "both", title = "Pie plot for chromosome")
}
\arguments{
\item{genefeatureall_cpgfeature_significantcpgisland}{refers to the input file with chromosomes, which can be files with/without significant filter().}

\item{genefeatureall_cpgfeature_significantshore}{refers to another input file with chromosomes, e.g. genefeatureall_cpgfeature_significantshore file for comparison, with default NULL.}

\item{methdirection}{refers to the methylation direction when the input file contains the methylation difference column i.e. Methdiff1 after Logic_regression(), 
which can be "hypo", "hyper" and "both", with the default "both" for both directions.}

\item{title}{refers to figure titles, with the default "Pie plot for chromosome".}
}
\value{
Outputs pie figure in different chromosomes.
}
\description{
This function outputs the pie plot for the percentages of sites or regions in different chromosomes.
}
\examples{
Chromosome_pieplot(genefeatureall_cpgfeature_significantcpgisland, title = "")
Chromosome_pieplot(genefeatureall_cpgfeature_significantcpgisland, title = "CpGisland")
Chromosome_pieplot(genefeatureall_cpgfeature_significantcpgisland, genefeatureall_cpgfeature_significantshore = genefeatureall_cpgfeature_significantshore, title = c("CpGisland","Shore"))
Chromosome_pieplot(siteall, title = "All cytosine sites") # Only consider the annotated chromosomes and the unannotated chromosomes will be discarded #
Chromosome_pieplot(siteall_Qvalue, title = "All cytosine sites")
Chromosome_pieplot(siteall_significant, title = "Significant cytosine sites")
Chromosome_pieplot(siteall_Qvalue, methdirection = "hyper", title = "Hyper-methylated distribution")
Chromosome_pieplot(siteall_significant, methdirection = "hypo", title = "Hypo-methylated pie plot")
Chromosome_pieplot(regiongeneall_Qvalue, methdirection = "hyper", title = "Hyper-methylated genes")

}
