% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circos_plot.R
\name{Circos_plot}
\alias{Circos_plot}
\title{Plot the circos}
\usage{
Circos_plot(inputcytofile, inputmethfile_QC, inputrefseqfile,
  inputcpgifeaturefile, labelname = regiongeneall_significant,
  linecolor = NULL)
}
\arguments{
\item{inputcytofile}{refers to the output of Cytofile_read() which contains the chromosome information.}

\item{inputmethfile_QC}{refers to input file with methylation levels after quality control.}

\item{inputrefseqfile}{refers to the output of Bedfile_read() which contains the gene information.}

\item{inputcpgifeaturefile}{refers to the output of Bedfile_read() which contains the CpG island and CpG island shore information.}

\item{labelname}{refers to the label of gene names which could be the significant genes after Significant_filter(), with default regiongeneall_significant with differentially methylated genes.
Sometimes, regiongenealls_significant will have some errors because it has unannotated chromosome name like chrUn_JH584304 or chrUn_NW_018084826v1. Thus, these chromosome names should be removed. 
If the labelname is from selfdefinedfile, then the file should contain the headers with chr (chromosome), start (start position), end (end position) and id (gene name).}

\item{linecolor}{refers to the colors of the lines plot for different methylation levels, with default NULL (black). 
If the linecolor is used, the length of colors should correspond to the length of groups.}
}
\value{
Outputs the circus figure with chromosomes, gene labels, the densities of the genes (track 3), CpG islands (track 4) and CpG island shores (track 5)
and the methylation levels of different groups from the outermost circle to the innermost circle.
}
\description{
This function outputs the circos plot for the methylation level and the density of gene, CpG island and CpG island shore on different chromosomes based on R RCircos package. 
All the files used in this function should contain chromosome, start position, and end position information that are required for R RCircos package.
}
\examples{
Circos_plot(inputcytofile, inputmethfile_QC, inputrefseqfile, inputcpgifeaturefile)
Circos_plot(inputcytofile, inputmethfile_QC, inputrefseqfile, inputcpgifeaturefile, 
            labelname = regiongeneall_significant, linecolor = c("blue1", "orange1", "green1"))
Circos_plot(inputcytofile, inputmethfile_QC, inputrefseqfile, inputcpgifeaturefile, 
            labelname = selfdefinedfile, linecolor = c("blue", "orange", "green"))

}
\references{
Hongen Zhang, Paul Meltzer, and Sean Davis. RCircos: an R package for Circos 2D track plots. BMC Bioinformatics, 2013, 14:244.
}
