% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box_plot.R
\name{Cpgfeature_boxplot}
\alias{Cpgfeature_boxplot}
\title{Internal Use Function.
That boxplots the methylation levels for one group based on CpG island features.}
\usage{
Cpgfeature_boxplot(
  genefeatureall_cpgfeature,
  groupnum = 1,
  ttest = TRUE,
  cpgfeaturelable = c("CpGisland", "Shore"),
  title = NULL,
  col = NULL,
  yaxislabel = TRUE
)
}
\arguments{
\item{genefeatureall_cpgfeature}{refers to output of Region_match as well. 
the output of gene or genebody match with cpgi features, e.g., all genes in promoter or exon or intron regions.}

\item{groupnum}{refers to the group number, can be NULL if without groups but the colnames should have "group" name 
like CpGislandgroup or Shoregroup, can also be automatically set in the Group_feature_boxplot function.}

\item{ttest}{refers to whether perform the Student t-test, with default TRUE.}

\item{cpgfeaturelable}{refers to CpG island features, with default CpGisland" and "Shore".}

\item{title}{refers to main text of tilte， with default "Group1".}

\item{col}{refers to colors, with default NULL.}

\item{yaxislabel}{refers to the label of y axis, with default TRUE,
but the yaxislabel will be set to FALSE when Group_feature_boxplot function is used for only one y axis in multiple groups.}
}
\value{
Outputs boxplot figure.
}
\description{
This function output the methylation levels in boxplot for one group based on CpG island features,
e.g., CpG island and CpG island shore based on groups.
}
\examples{
Cpgfeature_boxplot(genefeatureall_cpgfeature, groupnum = 1, ttest = TRUE, cpgfeaturelable = c("CpGisland", "Shore"), title = "Group1", col = "blue")

}
