% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_read.R
\name{Cytofile_read}
\alias{Cytofile_read}
\title{Read the cyto file.}
\usage{
Cytofile_read(
  paths = paste(system.file(package = "GeneDMRs"), "/methdata", sep = ""),
  cytofile = "cytoBandIdeo",
  suffix = ".txt.gz"
)
}
\arguments{
\item{paths}{refers to the path of input file, with default the package path.}

\item{cytofile}{refers to the name of input cyto file that is downloaded from UCSC website, with default "cytoBandIdeo".}

\item{suffix}{refers to the suffix of input cyto file, e.g., ".gz", ".zip" and so on (some files are in text .txt format, then ".txt" or ".txt.gz"), with default ".txt.gz".}
}
\value{
Outputs a data frame contains chromosome, start position, end position.
}
\description{
This function reads the chromosome information from cyto file (cytoBandIdeo.txt) and sort them by chromosome and position. 
The dataset of the example is the mouse genome information downloaded from UCSC website 
(http://hgdownload.cse.ucsc.edu/goldenPath/mm10/database/cytoBandIdeo.txt.gz).
}
\examples{
inputcytofile <- Cytofile_read()
inputcytofile <- Cytofile_read(paths = paste(system.file(package = "GeneDMRs"), "/methdata", sep=""), 
cytofile = "cytoBandIdeo", suffix = ".txt.gz")

}
