% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMR_test.R
\name{DMC_feature}
\alias{DMC_feature}
\title{Annotate the differentially methylated cytosine (DMC) to features.}
\usage{
DMC_feature(
  siteall_significant,
  featureid = TRUE,
  featurefile = list(inputgenebodyfile, inputcpgifeaturefile)
)
}
\arguments{
\item{siteall_significant}{refers to the input file with DMC sites.}

\item{featureid}{refers to whether to include the feature id or not, with the default TRUE. 
The feature id will output the id of first file of the featurefile list e.g., the id (gene name) of inputgenebodyfile.}

\item{featurefile}{refers to the input feature files e.g., inputgenebodyfile and inputcpgifeaturefile, 
with default two files in a list as featurefile = list(inputgenebodyfile, inputcpgifeaturefile),
and it can also be one file without a list e.g., featurefile = inputgenebodyfile.}
}
\value{
Outputs a data frame contains DMC sites with features.
}
\description{
This function annotates the differentially methylated cytosine (DMC) after statistical test Logic_regression() to gene body or CpG island features.
}
\examples{
siteall_significant_feature <- DMC_feature(siteall_significant, featurefile = list(inputgenebodyfile, inputcpgifeaturefile))
siteall_significant_feature <- DMC_feature(siteall_significant, featureid = FALSE, featurefile = list(inputgenebodyfile, inputcpgifeaturefile))
siteall_significant_feature <- DMC_feature(siteall_significant, featureid = TRUE, featurefile = inputgenebodyfile)

}
