% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methfile_QC.R
\name{DMC_methfile_QC}
\alias{DMC_methfile_QC}
\title{Merge the methylation file after quality control with DMCs}
\usage{
DMC_methfile_QC(inputmethfile_QC, siteall_significant)
}
\arguments{
\item{inputmethfile_QC}{refers to the input methylation file after quality control.}

\item{siteall_significant}{refers to the input DMCs file.}
}
\value{
Outputs a data frame by merging two input files of inputmethfile_QC and siteall_significant.
}
\description{
This function merges the methylation file after quality control of all samples with the DMCs after Significant_filter().
}
\examples{
DMC_inputmethfile_QC <- DMC_methfile_QC(inputmethfile_QC, siteall_significant)

}
