% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich_plot.R
\name{Enrich_plot}
\alias{Enrich_plot}
\title{Enrich plot for GO terms and pathways.}
\usage{
Enrich_plot(regiongenealls_significant, adjustpvaluecut = 0.1,
  enrichterm = "pathway", category = TRUE,
  Dbannotation = "org.Mm.eg.db", keggorganism = "mmu", listnum = 20,
  title = "Enrichment for significant gene",
  expressionfile_significant = NULL, expressionfile_genetype = NULL)
}
\arguments{
\item{regiongenealls_significant}{refers to the differentially methylated genes.}

\item{adjustpvaluecut}{refers to the threshold of the adjusted P values for the enrichment, with default 0.1.}

\item{enrichterm}{refers to the term need to be analyzed, which can be "GOgroup", "GO", "pathway", with default "pathway".}

\item{category}{refers to whether to divide the enrichments into two categories, i.e. hypo/hyper methylated or down/up regulated, with default TRUE.}

\item{Dbannotation}{refers to the annotation dataset, with default "org.Mm.eg.db" of mouse.}

\item{keggorganism}{refers to the species name for KEGG enrichment, with default "mmu" of mouse.}

\item{listnum}{refers to the list of display number, with default 20.}

\item{title}{refers to the title of figure, with default "Enrichments for significant gene".}

\item{expressionfile_significant}{refers to an additional file for differentially expressed genes which includes gene name and Log fold change (LogFC). 
This is a optional file for GO terms and pathways, with default NULL.}

\item{expressionfile_genetype}{refers to the gene type of expressionfile_significant file, which can be "REFSEQ", "ENTREZID", "SYMBOL",
or other gene types that can be used in clusterProfiler, with default NULL.}
}
\value{
Outputs a dot-plot figure of enrichment.
}
\description{
This function outputs the groups, GO terms and pathways plot for the enrichment based on R package clusterProfiler, org.Mm.eg.db (Mouse for example) and KEGG.db.
}
\examples{
Enrich_plot(regiongenealls_significant, enrichterm = "GOgroup", Dbannotation = "org.Mm.eg.db", title = "Biological process for significant gene")
Enrich_plot(regiongenealls_significant, enrichterm = "GO", title = "Go term for significant gene")
Enrich_plot(regiongenealls_significant, enrichterm = "GO", Dbannotation = "org.Hs.eg.db", title = "Go term for significant gene") # for human data #
Enrich_plot(regiongenealls_significant, adjustpvaluecut = 0.2, enrichterm = "pathway", keggorganism = "mmu", title = "Pathway for significant gene")
Enrich_plot(regiongenealls_significant, enrichterm = "pathway", keggorganism = "hsa", category = FALSE, title = "Pathway for significant gene") # for human data #

expressionfile_significant <- read.table(paste(system.file(package = "GeneDMRs"), "/methdata/DEgenes.txt", sep=""), header = T) # read DEgene file #

Enrich_plot(regiongenealls_significant, adjustpvaluecut = 0.2, enrichterm = "GO", Dbannotation = "org.Mm.eg.db", title = "Go term for significant gene in two categories", 
expressionfile_significant = expressionfile_significant, expressionfile_genetype = "SYMBOL")
Enrich_plot(regiongenealls_significant, enrichterm = "pathway", keggorganism = "mmu", title = "Pathway for significant gene in two categories", 
expressionfile_significant = expressionfile_significant, expressionfile_genetype = "SYMBOL")

}
