% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pie_plot.R
\name{Feature_pieplot}
\alias{Feature_pieplot}
\title{Pie plot based on different features}
\usage{
Feature_pieplot(siteall_significant_feature, methdirection = "both",
  title = "Pie plot for feature", threeDplot = TRUE)
}
\arguments{
\item{siteall_significant_feature}{refers to the input file with features, mainly for DMC sites with features.}

\item{methdirection}{refers to the methylation direction when the input file contains the methylation difference column i.e. Methdiff1 after Logic_regression(), 
which can be "hypo", "hyper" and "both", with the default "both" for both directions.}

\item{title}{refers to figure titles, with the default "Pie plot for chromosome".}

\item{threeDplot}{refers to whether to pie plot in three dimensions based on R pacakge plotrix, with the default TRUE.}
}
\value{
Outputs a pie figure in different features.
}
\description{
This function outputs the pie plot of feature percentages in gene body or CpG island mainly for DMC sites with features.
}
\examples{
Feature_pieplot(siteall_significant_feature)
Feature_pieplot(siteall_significant_feature, methdirection = "hypo")
Feature_pieplot(siteall_significant_feature, title = c("Gene body", "CpG island"))
Feature_pieplot(siteall_significant_feature, title = c("Pie plot for Gene body", "Pie plot for CpG island"), threeDplot = FALSE)
Feature_pieplot(siteall_significant_feature, methdirection = "hyper", title = c("Pie plot for Gene body", "Pie plot for CpG island"))

}
