% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box_plot.R
\name{Genebody_cpgfeature_boxplot}
\alias{Genebody_cpgfeature_boxplot}
\title{Boxplot the methylation levels for gene body based on CpG island features.}
\usage{
Genebody_cpgfeature_boxplot(genefeatureall_cpgfeature,
  genebodyname = c("promoters", "exons", "introns", "TSSes"),
  ttest = TRUE, cpgfeaturelable = c("CpGisland", "Shore"),
  title = c("Promoter", "Exon", "Intron", "TSS"), col = NULL)
}
\arguments{
\item{genefeatureall_cpgfeature}{refers to the input file of methylation levels with gene body and CpG island features.}

\item{genebodyname}{refers to the name of gene body features e.g. promoter, exon, intron and TSSes, with default "promoters", "exons", "introns", "TSSes".}

\item{ttest}{refers to whether perform the Student t-test, with default TRUE.}

\item{cpgfeaturelable}{refers to CpG island features, with default "CpGisland" and "Shore".}

\item{title}{refers to the figure title, with default "Promoter", "Exon", "Intron" and "TSS".}

\item{col}{refers to the boxplot colors, with default NULL.}
}
\value{
Outputs a boxplot figure with gene body and CpG island features.
}
\description{
This function outputs the methylation levels in boxplot for one or more features of gene body based on CpG island features, 
e.g. CpG island and CpG island shore features.
}
\examples{
Genebody_cpgfeature_boxplot(genefeatureall_cpgfeature)
Genebody_cpgfeature_boxplot(genefeatureall_cpgfeature, genebodyname = c("promoters","exons"), 
ttest = TRUE, cpgfeaturelable = c("CpGisland", "Shore"), title = c("Promoter", "Exon"), col = c("blue", "red"))

Genebody_cpgfeature_boxplot(genefeatureall_cpgfeature, genebodyname = c("promoters","exons","introns","TSSes"), 
ttest = TRUE, cpgfeaturelable = c("CpGisland", "Shore"), title = c("Promoters", "Exons", "Introns", "TSSes"), col = c("blue", "red", "green", "purple"))

}
