% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick_start.R
\name{GeneDMRs}
\alias{GeneDMRs}
\title{Gene-based differentially methylated regions analysis (GeneDMRs) and install the dependencies.}
\usage{
GeneDMRs(Dbannotation = "org.Mm.eg.db")
}
\arguments{
\item{Dbannotation}{refers to the annotation dataset for enrichment, with default "org.Mm.eg.db" of mouse.}
}
\value{
Outputs a list of required R packages.
}
\description{
GeneDMRs is an R package to detect the differentially methylated regions based on genes (DMG), gene body (DMP, DME, DMI), 
CpG islands and gene body interacted with CpG island features (e.g. DMG/DMP/DME/DMI_CpG island and DMG/DMP/DME/DMI_CpG island shore). 
This function can install the other R packages for the dependencies of GeneDMRs.
}
\examples{
GeneDMRs(Dbannotation = "org.Mm.eg.db")

}
