% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box_plot.R
\name{Group_boxplot}
\alias{Group_boxplot}
\title{Boxplot the methylation levels for groups.}
\usage{
Group_boxplot(regiongeneall, ttest = TRUE, title = "Group boxplot", col = NULL)
}
\arguments{
\item{regiongeneall}{refers to the input file with group methylation levels.}

\item{ttest}{refers to whether to perform the Student t-test, with default TRUE.}

\item{title}{refers to the figure title, with default "Group boxplot among genes".}

\item{col}{refers to the boxplot colors, with default NULL.}
}
\value{
Outputs boxplot figure with groups.
}
\description{
This function outputs the methylation levels of all the groups in boxplot without considering other features.
}
\examples{
Group_boxplot(regiongeneall)
Group_boxplot(genebodypromoterall, title = "Three groups among genes in promoter region")
Group_boxplot(regiongeneall, ttest = FALSE, title = "Three groups among genes", col = c("red", "green", "blue"))

}
