% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box_plot.R
\name{Group_cpgfeature_boxplot}
\alias{Group_cpgfeature_boxplot}
\title{Boxplot the methylation levels for groups based on CpG island features.}
\usage{
Group_cpgfeature_boxplot(
  genefeatureall_cpgfeature,
  groupnum = "all",
  ttest = TRUE,
  cpgfeaturelable = c("CpGisland", "Shore"),
  title = c("Group1", "Group2", "Group3"),
  col = NULL
)
}
\arguments{
\item{genefeatureall_cpgfeature}{refers to the input file with group methylation levels and CpG island features.}

\item{groupnum}{refers to the group number, with default "all" for all of the groups.}

\item{ttest}{refers to whether to perform the Student t-test, with default TRUE.}

\item{cpgfeaturelable}{refers to CpG island features, with default "CpGisland" and "Shore". Only one CpG island feature can also be available, e.g., "CpGisland".}

\item{title}{refers to the figure title， with default "Group1", "Group2" and "Group3".}

\item{col}{refers to the boxplot colors, with default NULL.}
}
\value{
Outputs a boxplot figure with groups and CpG island features.
}
\description{
This function outputs the methylation levels in boxplot for one or more groups based on CpG island features,
e.g., CpG island and CpG island shore features.
}
\examples{
Group_cpgfeature_boxplot(genefeatureall_cpgfeature, groupnum = 1)
Group_cpgfeature_boxplot(genefeatureall_cpgfeature, groupnum = "all", ttest = TRUE, cpgfeaturelable = c("CpGisland", "Shore"), 
title = c("Group1", "Group2", "Group3"), col = c("blue", "red", "green"))

}
