% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_plot.R
\name{Heatmap_plot}
\alias{Heatmap_plot}
\title{Heat map plot for chromosomes and features.}
\usage{
Heatmap_plot(regiongeneall_significant, featurename = NULL,
  title = "Methylation level", display_numbers = FALSE,
  number_format = "\%.0f", cluster_rows = FALSE, cluster_cols = TRUE,
  gaps_row = c(1, 2), gaps_col = NULL)
}
\arguments{
\item{regiongeneall_significant}{refers to the input file of methylation levels with differentially methylated genes or the genes in different gene body features.}

\item{featurename}{refers to the feature name of the output file from Significant_filter() for genefeatureall_cpgfeature file, that is "CpGisland" or "Shore", with default NULL.}

\item{title}{refers to the figure title, with the default "Methylation level".}

\item{display_numbers}{refers to whether to display the methylation value in the figure, with default FALSE.}

\item{number_format}{refers to the displayed number of the methylation value in round format.}

\item{cluster_rows}{refers to whether to cluster the row, with the default FALSE.}

\item{cluster_cols}{refers to whether to cluster the column, with the default TRUE.}

\item{gaps_row}{refers to whether to divide the row, with the default c(1,2) that divide the rows into three parts by row 1 and row 2.}

\item{gaps_col}{refers to whether to divide the column, with the NULL.}
}
\value{
Outputs a heat map figure with methylation levels.
}
\description{
This function outputs the heat map plot for methylation level in different chromosomes of differentially methylated genes with features based on R package pheatmap.
}
\examples{
Heatmap_plot(regiongeneall_significant)
Heatmap_plot(genefeatureall_cpgfeature_significantcpgisland, featurename = "CpGisland", display_numbers = FALSE, title = "Methylation level (\%) for genes with CpG island")
Heatmap_plot(genefeatureall_cpgfeature_significantshore, featurename = "Shore", title = "Methylation level (\%) for genes with shores")
Heatmap_plot(genefeatureall_cpgfeature_significantshore, featurename = "Shore", title = "Methylation level (\%) for genes with shores", cluster_cols = FALSE)
Heatmap_plot(regiongeneall_significant, title = "Methylation level (\%) for genes", display_numbers = FALSE)
Heatmap_plot(regiongeneall_significant, title = "Methylation level (\%) for genes", display_numbers = FALSE, cluster_rows = TRUE, gaps_row = NULL)

}
