% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMR_test.R
\name{Logic_regression}
\alias{Logic_regression}
\title{Logistical regression analysis for each region or each cytosine site.}
\usage{
Logic_regression(
  genefeatureall_cpgfeature,
  covariates = NULL,
  adjustedmethod = "fdr",
  diffgroup = NULL
)
}
\arguments{
\item{genefeatureall_cpgfeature}{refers to the input file with methylation levels to be tested.}

\item{covariates}{refers to the extra covariates used in the model, with the default NULL.}

\item{adjustedmethod}{refers to the methods to adjust P values to Q values, with the default "fdr" method.
The adjustedmethod could be "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "none" methods as well.}

\item{diffgroup}{refers to methylation difference between two groups, with the default NULL, that is the max group - min group.
The two groups can be manually selected e.g., diffgroup = c("group1", "group2").}
}
\value{
Outputs a data frame of region gene or region cpgi or those regions with different features or cytosine sites, 
by accompanying with P values, Q values and methylation differences.
}
\description{
This function tests each region or each cytosine site by logistical regression model to
achieve the P values and then be adjusted to Q values to account for multiple hypothesis testing.
}
\examples{
regiongeneall_Qvalue <- Logic_regression(regiongeneall)
regiongenealls_Qvalue <- Logic_regression(regiongenealls)
regioncpgall_Qvalue <- Logic_regression(regioncpgall , adjustedmethod = "fdr")
regiongenebodyall_Qvalue <- Logic_regression(regiongenebodyall, diffgroup = c("group1", "group2"))
regiongeneall_cpgfeature_Qvalue <- Logic_regression(regiongeneall_cpgfeature)
genefeatureall_cpgfeature_Qvalue <- Logic_regression(genefeatureall_cpgfeature)
genefeatureall_Qvalue <- Logic_regression(genefeatureall, adjustedmethod = "bonferroni")

siteall_Qvalue <- Logic_regression(siteall, adjustedmethod = "fdr") # for each cytosine site #
siteall_Qvalue <- Logic_regression(siteall, adjustedmethod = "fdr", diffgroup = c("group1", "group2"))

covariateinfo <- data.frame(Timepoint = c("week3","week3","week12","week12","week12"), Diet = c("Low","High","Low","High","High")) # for environmental factors #
regiongeneall_Qvalue <- Logic_regression(regiongeneall, covariates = covariateinfo)

}
