% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan_plot.R
\name{Manhattan_plot}
\alias{Manhattan_plot}
\title{Manhattan plot for all cytosines or regions.}
\usage{
Manhattan_plot(siteall_Qvalue, chrlabs = NULL, col = c("black",
  "grey"), ylab = "-log(Q value)", suggestiveline = -log10(0.01),
  genomewideline = -log10(0.001))
}
\arguments{
\item{siteall_Qvalue}{refers to the input file with Q value from DMR_test(), e.g. regiongeneall_Qvalues, genefeatureall_cpgfeature_Qvalue or others with Q values.}

\item{chrlabs}{refers to the label of chromosomes, with default NULL.}

\item{col}{refers to the color of plots, with default black and grey.}

\item{suggestiveline}{refers to the significant line, with default 0.01.}

\item{genomewideline}{refers to the genome-wide significant line, with default 0.001.}
}
\value{
Outputs a Manhattan figure with Q values.
}
\description{
This function outputs the Manhattan plot for all cytosines or regions in different chromosomes with significant line based on R package qqman.
}
\examples{
Manhattan_plot(siteall_Qvalue, ylab = "-log(Q-value)")
Manhattan_plot(regiongenealls_Qvalue, chrlabs = c(1:18,"X"), col = c("green","orange"), genomewideline = -log10(1e-02))
Manhattan_plot(genefeatureall_cpgfeature_Qvalue, ylab = c("-log(Q value) for CpG island", "-log(Q value) for Shore"), col = c("red","blue"),
suggestiveline = -log10(5e-02), genomewideline = -log10(1e-02))                        

}
