% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMR_test.R
\name{Meth_difference}
\alias{Meth_difference}
\title{Internal Use Function
That calculates the methylation difference between groups.}
\usage{
Meth_difference(output, diffgroup = NULL)
}
\arguments{
\item{output}{refers to the output of Methmean_match.}

\item{diffgroup}{refers to methylation difference between two groups, with the default NULL that is the max group - min group.
The two groups can be manually select like diffgroup = c("group1", "group2").}
}
\value{
Outputs methylation difference between groups.
}
\description{
This function calculate the methylation difference between groups of each region or each cytosine.
}
\examples{
output_methdiff <- Meth_difference(output)
adjustedoutput_methdiff <- Meth_difference(adjustedoutput, diffgroup = c("group1", "group2"))

}
