% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methmean_match.R
\name{Meth_mean}
\alias{Meth_mean}
\title{Internal Use Function}
\usage{
Meth_mean(inputmethfile_QC, regionchr, cpgifeaturefile = NULL)
}
\arguments{
\item{inputmethfile_QC}{refers to the inputmethfile after quality control.}

\item{regionchr}{refers to the filtered subset of inputrefseqfile or inputcpgifile.}

\item{cpgifeaturefile}{refers to the whether have inputcpgifeaturefile or not.}
}
\value{
Outputs the mean methylation.
}
\description{
This function calculates the mean methylation difference along each read gene.
Formula:The mean methylation of a gene of one treatment group was calculated by: 
¡Æn((¡Æm[MR]_ij )/(¡Æm[TR]_ij ))*W_ij and W_ij =  (¡Æm[TR]_ij )/(¡Æn¡Æm[TR]_ij ), 
where [MR]_ij and [TR]_ij are methylated and total reads number of the involved CpG/DMC j at a given gene of individual i, 
n is the total individual number of one treatment group, m is total number of CpG/DMC involved in this gene and 
W_ij is the weight of reads of the involved CpG/DMC j of individual i. 
Firstly calculate methylation difference through reads of all cytosines along gene 
Secondly calculate methylation difference along each group
When provide inputcpgifeaturefile to calculate the gene mean methylation difference ,
two features will be combined that CpG island and CpG island shore.
}
\examples{
Meth_mean(inputmethfile_QC, regionchr)
Meth_mean(inputmethfile_QC, regionchr, cpgifeaturefile = inputcpgifeaturefile)

}
