% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick_start.R
\name{Prepare_GeneDMRs}
\alias{Prepare_GeneDMRs}
\title{Install the dependencies.}
\usage{
Prepare_GeneDMRs(Dbannotation = "org.Mm.eg.db")
}
\arguments{
\item{Dbannotation}{refers to the annotation dataset for enrichment, with default "org.Mm.eg.db" of mouse.}
}
\value{
Outputs a list of installed R packages.
}
\description{
This function install the other R packages for the dependencies of GeneDMRs.
}
\examples{
Prepare_GeneDMRs(Dbannotation = "org.Mm.eg.db")

}
