% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMR_test.R
\name{Pvalue_adjusted}
\alias{Pvalue_adjusted}
\title{Internal Use Function
Adjust the P values of logistical regression model.}
\usage{
Pvalue_adjusted(output, adjustedmethod = "fdr")
}
\arguments{
\item{output}{refers to the output of Methmean_match.}

\item{adjustedmethod}{refers to the P values adjusted methods, with the default fdr.
The adjustedmethod could be "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "none" as well.}
}
\value{
Outputs the regiongene or regioncpg or of gene/cpg with different features with Q values.
}
\description{
This function adjust the P values of each region or each cytosine by logistical regression model.
}
\examples{
output_Qvalue <- Pvalue_adjusted(output, adjustedmethod = "fdr")
output_Qvalue <- Pvalue_adjusted(output)
output_Qvalue <- Pvalue_adjusted(output, adjustedmethod = "bonferroni")
output_Qvalue <- Pvalue_adjusted(output, adjustedmethod = "BH")

}
