% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick_start.R
\name{Quick_DMCs}
\alias{Quick_DMCs}
\title{Quick use the GeneDMRs package for differentially methylated cytosine sites.}
\usage{
Quick_DMCs(
  paths = paste(system.file(package = "GeneDMRs"), "/methdata", sep = ""),
  suffixmeth = ".gz",
  control_paths = NULL,
  case_paths = NULL
)
}
\arguments{
\item{paths}{refers to the path of input file, with default the package path.}

\item{suffixmeth}{refers to the suffix of methylation file, e.g., ".gz", ".zip" and so on (some files are in text .txt format, then ".txt" or ".txt.gz"), with default ".gz".}

\item{control_paths}{refers to the path of control groups, with default NULL.}

\item{case_paths}{refers to the path of case groups, with default NULL.}
}
\value{
Outputs DMC results.
}
\description{
This function outputs the differentially methylated cytosine sites (DMCs).
}
\examples{
allDMCs <- Quick_DMCs()

# if only case and control group (n = 2) paths are provided #
controls <- c("C:/Users/GeneDMRs/methdata/1_1.gz", "C:/Users/GeneDMRs/methdata/1_2.gz", "C:/Users/GeneDMRs/methdata/1_3.gz")
cases <- c("C:/Users/GeneDMRs/methdata/2_1.gz", "C:/Users/GeneDMRs/methdata/2_1.gz")
allDMCs <- Quick_DMCs(control_paths = controls, case_paths = cases)

}
