% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_read.R
\name{refseqfile_check}
\alias{refseqfile_check}
\title{Internal Use Function
That checks and annotates the promoter to the specific gene.}
\usage{
refseqfile_check(geneobj, regionfile)
}
\arguments{
\item{geneobj}{refers to the genebody file from readTranscriptFeatures of package genomation.}

\item{regionfile}{refers to the reference gene or cpg bed file (without header)}
}
\value{
Outputs gene body file.
}
\description{
This function sorts and makes sure geneobj file can be used in this package,
that mainly works on the promoters without the choromsome and position information.
In addition, it can add the genomic number of exons and introns, e.g., first exon or second intron.
}
\examples{
geneobj <- refseqfile_check(geneobj, regionfile)

}
