% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_read.R
\name{Regionfile_read}
\alias{Regionfile_read}
\title{Internal Use Function
That reads the bed file.}
\usage{
Regionfile_read(bedfile, suffix)
}
\arguments{
\item{bedfile}{refers to the bedfile name (e.g., refseq or cpgi) downloaded from UCSC, with the default refseq.}

\item{suffix}{refers to the compressed file suffix such as ".gz", ".zip" and so on, with the default suffix ".gz".}
}
\value{
Outputs bed file that can be used in the next step.
}
\description{
This function reads the bed file and sort them.
}
\examples{
regionfile <- Regionfile_read(bedfile, suffix)

}
