% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box_plot.R
\name{Sample_boxplot}
\alias{Sample_boxplot}
\title{Boxplot the methylation levels or read numbers in different samples.}
\usage{
Sample_boxplot(
  inputmethfile,
  inputrefseqfile,
  Meth_plot = TRUE,
  ylab = "Methylation level",
  refseqname = NULL,
  col = NULL
)
}
\arguments{
\item{inputmethfile}{refers to the input file with methylation levels.}

\item{inputrefseqfile}{refers to the input of gene regions.}

\item{Meth_plot}{refers to whether to plot the methylation levels, with default TRUE, otherwise to plot the read numbers.}

\item{ylab}{refers to the label of y axis, with default "Methylation level".}

\item{refseqname}{refers to NCBI ID of specific gene, with default NULL.}

\item{col}{refers to boxplot colors, with default NULL.}
}
\value{
Outputs a boxplot figure with all the samples.
}
\description{
This function outputs the methylation levels or read numbers of the selected genes or all the genes in the different samples.
}
\examples{
Sample_boxplot(inputmethfile_QC, inputrefseqfile)
Sample_boxplot(inputmethfile_QC, inputrefseqfile, refseqname = "NM_001244864")
Sample_boxplot(inputmethfile_QC, inputrefseqfile, refseqname = c("NM_001244864", "NM_001244534"))
Sample_boxplot(inputmethfile_QC, inputrefseqfile, , ylab = "Methylation level (\%)"， refseqname = c("NM_001244864", "NM_001143697", "NM_213902"), col = c("red", "green", "blue"))
Sample_boxplot(inputmethfile_QC, inputrefseqfile, Meth_plot = FALSE, ylab = "Read number", col = c("red", "blue"))
Sample_boxplot(inputmethfile_QC, inputrefseqfile, Meth_plot = FALSE, ylab = "Read number", refseqname = c("NM_001244864", "NM_001244534"))
Sample_boxplot(inputmethfile_QC, inputrefseqfile, Meth_plot = FALSE, ylab = "Read number", 
refseqname = c("NM_001244864", "NM_001143697", "NM_213902"), col = c("red", "green", "blue"))

}
