% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMR_test.R
\name{Significant_filter}
\alias{Significant_filter}
\title{Filter the significant regions or cytosine sites.}
\usage{
Significant_filter(genefeatureall_cpgfeature_Qvalue, qvalue = 0.01,
  methdiff = 0, featureout = 1)
}
\arguments{
\item{genefeatureall_cpgfeature_Qvalue}{refers to the input file with Q values and methylation differences need to be filtered.}

\item{qvalue}{refers to the threshold of Q values that Q values larger than this will be discarded, with default 0.01.}

\item{methdiff}{refers to the threshold of methylation differences that methylation differences less than this will be discarded, with the default 0.}

\item{featureout}{refers to which feature will be filtered, with default 1. When featureout = 2, it means that the second feature will be filtered and outputted.}
}
\value{
Outputs a data frame of the significant regions or cytosine sites.
}
\description{
This function filters significant regions or cytosine sites based on Q value and methylation difference.
}
\examples{
genefeatureall_cpgfeature_significantcpgisland <- Significant_filter(genefeatureall_cpgfeature_Qvalue)
genefeatureall_cpgfeature_significantshore <- Significant_filter(genefeatureall_cpgfeature_Qvalue, featureout = 2)
genefeatureall_cpgfeature_significantcpgisland <- Significant_filter(genefeatureall_cpgfeature_Qvalue, qvalue = 0.001, methdiff = 0.01, featureout = 1)
regiongeneall_significant <- Significant_filter(regiongeneall_Qvalue)
regiongenealls_significant <- Significant_filter(regiongenealls_Qvalue, methdiff = 0.05)
siteall_significant <- Significant_filter(siteall_Qvalue)
siteall_significant <- Significant_filter(siteall_Qvalue, qvalue = 0.001, methdiff = 0.1)

}
