% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venn_plot.R
\name{Venn_plot}
\alias{Venn_plot}
\title{Venn plot for the common CpG island and CpG island shore.}
\usage{
Venn_plot(
  genefeatureall_cpgfeature_Qvalue,
  title = "Venn plot",
  fillcolor = c("cornflowerblue", "green")
)
}
\arguments{
\item{title}{refers to the figure title, with default "Venn plot".}

\item{fillcolor}{refers to the filled color, with default "cornflowerblue"and "green".}

\item{genefeatureall_cpgfeature}{refers to the input file with Q values of two features.}
}
\value{
Outputs a venn figure in two features.
}
\description{
This function outputs the venn plot for the common CpG island and CpG island shore regions 
that are covered by methylated cytosine sites based on R package VennDiagram.
}
\examples{
Venn_plot(genefeatureall_cpgfeature)
Venn_plot(genefeatureall_cpgfeature_Qvalue)
Venn_plot(genefeatureall_cpgfeature_Qvalue, fillcolor = c("red","blue"))

}
