% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan_plot.R
\name{Volcano_plot}
\alias{Volcano_plot}
\title{Volcano plot for all the cytosines.}
\usage{
Volcano_plot(siteall_Qvalue,
  title = "Volcano for Q value and methylation difference",
  qvalue = 0.01, methdiffpercentage = c(5, 10, 15, 20, 25),
  pointcolor = c("red", "purple", "orange", "yellow", "blue", "green"))
}
\arguments{
\item{siteall_Qvalue}{refers to the input file with Q values and methylation differences.}

\item{title}{refers to the figure title, with default "Volcano for Q value and methylation difference".}

\item{qvalue}{refers to the threshold of Q values that Q values less than this will be colored, with default 0.01.}

\item{methdiffpercentage}{refers to the threshold of methylation level (%) differences that 
methylation differences larger than this will be colored, with default 5, 10, 15, 20, 25.}

\item{pointcolor}{refers to the point plot color, with default "red", "purple", "orange", "yellow", "blue", "green".}
}
\value{
Outputs a volcano figure.
}
\description{
This function outputs the volcano plot for all the cytosines with Q values and methylation differences.
}
\examples{
Volcano_plot(siteall_Qvalue)
Volcano_plot(siteall_Qvalue, pointcolor = c("red", "blue", "yellow", "purple", "orange", "green"))
Volcano_plot(siteall_Qvalue, title = "Volcano plot", qvalue = 0.001, methdiffpercentage = c(10, 15, 20, 30, 40), 
pointcolor = c("red", "purple", "orange", "yellow", "blue", "green"))

}
