% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Window_divide.R
\name{Window_divide}
\alias{Window_divide}
\title{Divide the genome to windows}
\usage{
Window_divide(inputcytofile, windowbp = 1e+06)
}
\arguments{
\item{inputcytofile}{refers to the input cyto file with chromosome information.}

\item{windowbp}{refers to the window length in base pair (bp) to be divided, with default 1,000,000.}
}
\value{
Outputs a data frame with window regions.
}
\description{
This function outputs the window regions of the whole genome.
}
\examples{
windowfile <- Window_divide(inputcytofile)
windowfile <- Window_divide(inputcytofile, windowbp = 10000)

}
