% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pie_plot.R
\name{chromosmoe_sort}
\alias{chromosmoe_sort}
\title{Internal Use Function
That sorts the chromosomes from "chr1", "chr2", "chr3"......, "chrX" and "chrY".}
\usage{
chromosmoe_sort(chromtable)
}
\arguments{
\item{chromtable}{refers to the input file for sorting the chromosomes.}
}
\value{
Outputs the sorted file by chromosome number.
}
\description{
This function sorts the chromosomes.
}
\examples{
chromtable <- chromosmoe_sort(chromtable)

}
