% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methfile_QC.R
\name{Methfile_QC}
\alias{Methfile_QC}
\title{Quality control for the input methylation file.}
\usage{
Methfile_QC(inputmethfile, low_coveragenum = 10,
  high_coveragenum = NULL, low_quantile = NULL, high_quantile = 99.9,
  coveragewrite = TRUE)
}
\arguments{
\item{inputmethfile}{refers to the input of methylation file after Methfile_read().}

\item{low_coveragenum}{refers to the minimum read coverage to be discarded, with default 10.}

\item{high_coveragenum}{refers to the maximum read coverage to be discarded, with default NULL.}

\item{low_quantile}{refers to the minimum quantile of read coverage to be discarded, with default NULL.}

\item{high_quantile}{refers to the maximum quantile of read coverage to be discarded, with default 99.99.}

\item{coveragewrite}{refers to whether to write out the read coverage file to the given path, with default TRUE.}
}
\value{
Outputs a data frame contain chromosome, position, and Cs & Ts for different replicates and groups after quality control.
}
\description{
This function discards the cytosine sites with low read coverage (quantile) or high read coverage (quantile).
}
\examples{
inputmethfile_QC <- Methfile_QC(inputmethfile)
inputmethfile_QC <- Methfile_QC(inputmethfile, low_coveragenum = 20, high_quantile = 99.99)
inputmethfile_QC <- Methfile_QC(inputmethfile, low_coveragenum = 10, high_coveragenum = 100, coveragewrite = FALSE)

}
