% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick_start.R
\name{Quick_DMCs}
\alias{Quick_DMCs}
\title{Quick use the GeneDMRs package for differentially methylated cytosine sites}
\usage{
Quick_DMCs(paths = paste(system.file(package = "GeneDMRs"), "/methdata",
  sep = ""), suffixmeth = ".gz")
}
\arguments{
\item{paths}{refers to the path of input file, with default the package path.}

\item{suffixmeth}{refers to the suffix of methylation file, e.g. ".gz", ".zip" and so on (some files are in text .txt format, then ".txt" or ".txt.gz"), with default ".gz".}
}
\value{
Outputs DMC results.
}
\description{
This function outputs the differentially methylated cytosine sites (DMCs).
}
\examples{
Quick_DMCs()

}
