% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick_start.R
\name{Quick_GeneDMRs}
\alias{Quick_GeneDMRs}
\title{Quick use the GeneDMRs package for gene based differentially methlated regions.}
\usage{
Quick_GeneDMRs(
  paths = paste(system.file(package = "GeneDMRs"), "/methdata", sep = ""),
  suffixmeth = ".gz",
  bedfile = "refseq",
  suffixbed = ".txt",
  Dbannotation = "org.Mm.eg.db",
  keggorganism = "mmu"
)
}
\arguments{
\item{paths}{refers to the path of input file, with default the package path.}

\item{suffixmeth}{refers to the suffix of methylation file, e.g., ".gz", ".zip" and so on (some files are in text .txt format, then ".txt" or ".txt.gz"), with default ".gz".}

\item{bedfile}{refers to the file name of bed file for "refseq". This file is downloaded from UCSC website, with default "refseq".}

\item{suffixbed}{refers to the suffix of bed file, e.g., ".gz", ".zip" and so on (some files are in text .txt format, then ".txt" or ".txt.gz"), with default ".txt".}

\item{Dbannotation}{refers to the annotation dataset for enrichment, with default "org.Mm.eg.db" of mouse.}

\item{keggorganism}{refers to the species name for KEGG enrichment, with default "mmu" of mouse.}
}
\value{
Outputs a series of DMG results.
}
\description{
This function outputs a series of results and figures for gene based regions' methylation analysis.
}
\examples{
allDMGs <- Quick_GeneDMRs()
allDMGs_mouse <- Quick_GeneDMRs(Dbannotation = "org.Mm.eg.db", keggorganism = "mmu")

}
