% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_read.R
\name{refseqfile_check}
\alias{refseqfile_check}
\title{Internal Use Function}
\usage{
refseqfile_check(geneobj, regionfile)
}
\arguments{
\item{geneobj}{refers to the genebody file from readTranscriptFeatures of package genomation.}

\item{regionfile}{refers to the reference gene or cpg bed file (without header)}
}
\value{
Outputs gene body file.
}
\description{
This function sorts and makes sure geneobj file can be used in this package,
that mainly works on the promoters without the choromsome and position information.
}
\examples{
geneobj <- refseqfile_check(geneobj, regionfile)

}
